pyproject.toml
[build-system]
requires = ["setuptools>=61", "setuptools-scm"]
build-backend = "setuptools.backends.legacy:build"

[project]
name = "opv-fair"
version = "0.1.0"
description = "FAIR pipeline for OPV degradation IV measurements (L1 + DC systems)"
readme = "README.md"
license = { text = "MIT" }
requires-python = ">=3.10"
dependencies = [
    "nomad-lab>=1.3.0",
    "pandas>=2.0",
    "numpy>=1.24",
    "h5py>=3.8",
    "openpyxl>=3.1",
]

[project.optional-dependencies]
dev = ["pytest", "ruff", "mypy"]

# NOMAD plugin entry points
[project.entry-points."nomad.plugin"]
opv_schema  = "opv_fair.schema:m_package"
opv_parser_l1 = "opv_fair.parsers.l1_parser:m_parser"
opv_parser_dc = "opv_fair.parsers.dc_parser:m_parser"

[tool.setuptools.packages.find]
where = ["src"]

[tool.ruff]
line-length = 100
